+++ src/test/java/org/delia/bdd/AllBDDTests.java
@@ -148,7 +148,6 @@ public class AllBDDTests extends NewBDDBase {
 	@Test
 	public void testR1500() {
 		runR1500File("t0-queryfn-orderby.txt", 4);
-		runR1500File("t0-queryfn-orderby-2span.txt", 1);
 		runR1500File("t0-queryfn-distinct.txt", 3);
 		runR1500File("t0-queryfn-distinct-relation.txt", 2);
 		runR1500File("t0-queryfn-flatten.txt", 0);
@@ -178,6 +177,11 @@ public class AllBDDTests extends NewBDDBase {
 		runR1500File("t0-queryfn-avg.txt", 0);
 		runR1500File("t0-queryfn-limit.txt", 5);
 		runR1500File("t0-queryfn-offset.txt", 5);
+	}
+	
+	@Test
+	public void testR1500a() {
+		runR1500File("t0-queryfn-orderby-2span.txt", 1);
 	}
 	
 	@Test
