+++ src/test/java/org/delia/parser/SchemaGeneratorTests.java
@@ -2,10 +2,10 @@ package org.delia.parser;
 
 import static org.junit.Assert.assertEquals;
 
+import org.delia.base.FakeTypeCreator;
 import org.delia.db.schema.SchemaFingerprintGenerator;
 import org.delia.runner.TypeRegistryHelper;
 import org.delia.type.DTypeRegistry;
-import org.delia.typebuilder.FakeTypeCreator;
 import org.junit.Test;
