+++ src/test/java/org/delia/runner/ValidationRuleRunnerTests.java
@@ -2,6 +2,7 @@ package org.delia.runner;
 
 import static org.junit.Assert.assertEquals;
 
+import org.delia.base.FakeTypeCreator;
 import org.delia.db.DBAccessContext;
 import org.delia.db.DBExecutor;
 import org.delia.rule.AlwaysRuleGuard;
@@ -13,7 +14,6 @@ import org.delia.rule.rules.ContainsRule;
 import org.delia.type.DType;
 import org.delia.type.DTypeRegistry;
 import org.delia.type.DValue;
-import org.delia.typebuilder.FakeTypeCreator;
 import org.delia.valuebuilder.StringValueBuilder;
 import org.junit.Before;
 import org.junit.Test;
